backend/requirements.txt
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
coverage==6.3.2
crispy-bootstrap5==0.6
Django==4.0.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-filter==21.1
django-js-asset==2.0.0
django-taggit==2.1.0
djangorestframework==3.13.1
importlib-metadata==4.11.2
iniconfig==1.1.1
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
sqlparse==0.4.2
tomli==2.0.1
zipp==3.7.0
djangorestframework-simplejwt